Refactor: ApiError generalError파일로 병합
