--- server/utils/ApiError.js
@@ -1,15 +1,0 @@
-class ApiError extends Error {
-  constructor(statusCode, name, message, isOperational = true, stack = "") {
-    super(message);
-    this.name = name;
-    this.statusCode = statusCode;
-    this.isOperational = isOperational;
-    if (stack) {
-      this.stack = stack;
-    } else {
-      Error.captureStackTrace(this, this.constructor);
-    }
-  }
-}
-
-module.exports = ApiError;
