+++ server/utils/generalError.js
@@ -1,5 +1,18 @@
-const ApiError = require("./ApiError");
 const httpStatus = require("http-status");
+
+class ApiError extends Error {
+  constructor(statusCode, name, message, isOperational = true, stack = "") {
+    super(message);
+    this.name = name;
+    this.statusCode = statusCode;
+    this.isOperational = isOperational;
+    if (stack) {
+      this.stack = stack;
+    } else {
+      Error.captureStackTrace(this, this.constructor);
+    }
+  }
+}
 
 class BadRequestError extends ApiError {
   constructor(message) {
@@ -38,6 +51,7 @@ class InternalServerError extends ApiError {
 }
 
 module.exports = {
+  ApiError,
   BadRequestError,
   UnauthorizedError,
   ForbiddenError,
