+++ server/utils/index.js
@@ -1,4 +1,3 @@
-const ApiError = require("./ApiError");
 const catchAsync = require("./catchAsync");
 const generalError = require("./generalError");
 const parse = require("./parse");
@@ -6,15 +5,16 @@ const parse = require("./parse");
 const s3Upload = require("./s3Upload");
 
 module.exports = {
-  ApiError,
   catchAsync,
+  ApiError: generalError.ApiError,
   BadRequestError: generalError.BadRequestError,
   UnauthorizedError: generalError.UnauthorizedError,
   ForbiddenError: generalError.ForbiddenError,
   NotFoundError: generalError.NotFoundError,
   InternelServerError: generalError.InternalServerError,
   uploadProdImg: s3Upload.uploadProductImg,
-  parseProducts: parse.parseProducts,
   deleteProdImg: s3Upload.deleteProductImg,
   getImgUrlByKey: s3Upload.getImgUrlByKey,
+  parseProducts: parse.parseProducts,
+  getProductsQueryConvert: parse.getProductsQueryConvert,
 };
